# apps/api/openapi.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"appGetData","parameters":[],"responses":{"200":{"description":"Application data retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppDataResponseDto"}}}}},"summary":"Get application data","tags":["app"]}},"/api/auth/register":{"post":{"operationId":"authRegister","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponseDto"}}}},"400":{"description":"Bad request - validation error"}},"summary":"Register a new user","tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"authLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}},"401":{"description":"Invalid credentials"}},"summary":"Login user","tags":["auth"]}},"/api/auth/logout":{"post":{"operationId":"authLogout","parameters":[],"responses":{"200":{"description":"Logout successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutResponseDto"}}}}},"summary":"Logout user","tags":["auth"]}},"/api/auth/register-admin":{"post":{"operationId":"authRegisterAdmin","parameters":[],"responses":{"201":{"description":"Admin user registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponseDto"}}}},"400":{"description":"Bad request - validation error"}},"summary":"Register a new admin user","tags":["auth"]}},"/api/todos":{"post":{"operationId":"todosCreate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"201":{"description":"Todo created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoResponseDto"}}}},"401":{"description":"Unauthorized"}},"security":[{"bearer":[]}],"summary":"Create a new todo","tags":["todos"]},"get":{"operationId":"todosFindAll","parameters":[],"responses":{"200":{"description":"Todos retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoResponseDto"}}}}},"401":{"description":"Unauthorized"}},"security":[{"bearer":[]}],"summary":"Get all todos for the current user","tags":["todos"]}},"/api/todos/admin/all":{"get":{"operationId":"todosFindAllForAdmin","parameters":[],"responses":{"200":{"description":"All todos retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoWithProfileResponseDto"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden - Admin role required"}},"security":[{"bearer":[]}],"summary":"Get all todos (admin only)","tags":["todos"]}},"/api/todos/{id}":{"get":{"operationId":"todosFindOne","parameters":[{"name":"id","required":true,"in":"path","description":"Todo ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Todo retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoWithProfileResponseDto"}}}},"401":{"description":"Unauthorized"},"404":{"description":"Todo not found"}},"security":[{"bearer":[]}],"summary":"Get a specific todo by ID","tags":["todos"]},"patch":{"operationId":"todosUpdate","parameters":[{"name":"id","required":true,"in":"path","description":"Todo ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoDto"}}}},"responses":{"200":{"description":"Todo updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoResponseDto"}}}},"401":{"description":"Unauthorized"},"404":{"description":"Todo not found"}},"security":[{"bearer":[]}],"summary":"Update a todo","tags":["todos"]},"delete":{"operationId":"todosRemove","parameters":[{"name":"id","required":true,"in":"path","description":"Todo ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Todo deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteTodoResponseDto"}}}},"401":{"description":"Unauthorized"},"404":{"description":"Todo not found"}},"security":[{"bearer":[]}],"summary":"Delete a todo","tags":["todos"]}}},"info":{"title":"Todo API","description":"The Todo API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AppDataResponseDto":{"type":"object","properties":{"message":{"type":"string","description":"Application message","example":"Hello API"}},"required":["message"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","description":"User email","example":"user@example.com"},"password":{"type":"string","description":"User password","example":"SecurePassword123!"},"isAdmin":{"type":"boolean","description":"Whether to register as admin","default":false}},"required":["email","password"]},"RegisterResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"User ID","example":"550e8400-e29b-41d4-a716-446655440000"},"email":{"type":"string","description":"User email","example":"user@example.com"},"role":{"type":"string","description":"User role","example":"user"}},"required":["id","email","role"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"User email","example":"user@example.com"},"password":{"type":"string","description":"User password","example":"SecurePassword123!"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"access_token":{"type":"string","description":"JWT access token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}},"required":["access_token"]},"LogoutResponseDto":{"type":"object","properties":{"message":{"type":"string","description":"Success message","example":"Logout successful"}},"required":["message"]},"CreateTodoDto":{"type":"object","properties":{"title":{"type":"string","description":"The title of the todo","example":"Buy groceries"},"description":{"type":"string","description":"Optional description of the todo","example":"Buy milk, eggs, and bread"},"completed":{"type":"boolean","description":"Whether the todo is completed","default":false}},"required":["title"]},"TodoResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Todo ID","example":"550e8400-e29b-41d4-a716-446655440000"},"title":{"type":"string","description":"The title of the todo","example":"Buy groceries"},"description":{"type":"string","description":"Description of the todo","example":"Buy milk, eggs, and bread"},"completed":{"type":"boolean","description":"Whether the todo is completed","example":false},"user_id":{"type":"string","description":"User ID who owns this todo","example":"550e8400-e29b-41d4-a716-446655440001"},"created_at":{"type":"string","description":"When the todo was created","example":"2024-01-01T00:00:00.000Z"},"updated_at":{"type":"string","description":"When the todo was last updated","example":"2024-01-01T00:00:00.000Z"}},"required":["id","title","completed","user_id","created_at","updated_at"]},"TodoWithProfileResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Todo ID","example":"550e8400-e29b-41d4-a716-446655440000"},"title":{"type":"string","description":"The title of the todo","example":"Buy groceries"},"description":{"type":"string","description":"Description of the todo","example":"Buy milk, eggs, and bread"},"completed":{"type":"boolean","description":"Whether the todo is completed","example":false},"user_id":{"type":"string","description":"User ID who owns this todo","example":"550e8400-e29b-41d4-a716-446655440001"},"created_at":{"type":"string","description":"When the todo was created","example":"2024-01-01T00:00:00.000Z"},"updated_at":{"type":"string","description":"When the todo was last updated","example":"2024-01-01T00:00:00.000Z"},"profiles":{"type":"object","description":"Profile information of the todo owner","properties":{"email":{"type":"string","example":"user@example.com"}}}},"required":["id","title","completed","user_id","created_at","updated_at","profiles"]},"UpdateTodoDto":{"type":"object","properties":{"title":{"type":"string","description":"The title of the todo","example":"Buy groceries"},"description":{"type":"string","description":"Description of the todo","example":"Buy milk, eggs, and bread"},"completed":{"type":"boolean","description":"Whether the todo is completed"}}},"DeleteTodoResponseDto":{"type":"object","properties":{"message":{"type":"string","description":"Success message","example":"Todo deleted successfully"}},"required":["message"]}}}}